requirements.txt
flake8==3.8.4
Flask==2.0.2
Flask-APScheduler==1.12.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Mail==0.9.0
flask-marshmallow==0.14.0
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
gunicorn==20.0.4
Werkzeug==2.0.1
marshmallow-sqlalchemy==0.24.1
python-dotenv==0.15.0
psycopg2==2.8
pytest==6.0.2
pytest-cov==2.10.1
Flask-Bcrypt==0.7.1
numpy==1.18.0